test(examples): run the e2e suite against both sqlite and postgres
Extract the CRUD e2e suite's assertions into crud-e2e.suite.ts as
registerCrudE2eSuite(getApp), parameterized instead of closing over a
module-level app. app.e2e.spec.ts (sqlite, unchanged behavior) and the new
app-postgres.e2e.spec.ts each boot the app against their own database and
call the same suite — one behavioral spec, two drivers, no forked
assertions.

app-postgres.e2e.spec.ts self-provisions Postgres via Testcontainers in
beforeAll, so pnpm check exercises both databases with no manual setup
(it does require a running Docker daemon).
